git-svn-id: http://svn.bluexml.com/private/S-IDE@28145 c4e9bb49-730a-0410-b52d-bc9a0859fdfe

+++ Integration/trunk/com.bluexml.side.Integration.buildHudson/config/launch_maven_subProcess.sh
@@ -11,42 +11,45 @@ else
   exit -2
 fi
 
-
-cp $WORKSPACE/../buildAuto/Ankle/delete_folder.pl $WORKSPACE/../buildAuto/Ankle/$WORKING_DIR
-cd $WORKSPACE/../buildAuto/Ankle/$WORKING_DIR
-perl delete_folder.pl pom.xml
-cp $WORKSPACE/superpom/pom.xml .
-
-echo "update superpom"
-perl -pi -le 'print "<modules>" if $. == 12' pom.xml
-perl -pi -le 'print "</modules>" if $. == 13' pom.xml
-for s in `ls -d */ `; do
-    repertoire=$s
-	name=`perl -e '"'$repertoire'"=~/([a-zA-Z0-9._]*)/; print $1';`
-  	res="<module>$name</module>"
-  	echo $repertoire $name $res
-	perl -pi -le 'print "'$res'" if $. == 13' pom.xml
-done
-echo "superpom patched"
-echo "Maven deploy start"
-mvn clean deploy -e -P public > log_maven.log
-echo "Maven deploy done"
-
-sleep 5m
-
-tail -100 log_maven.log
-
-if grep "BUILD SUCCESSFUL" log_maven.log ; then
-	echo "Maven deploy success"
-else
-
-	if grep "ERROR" log_maven.log ; then	
-		echo "Maven deploy failed"
-		exit 1
-	else
-		echo "Maven deploy no error"
-	fi
+if [ -f delete_folder.pl ]; then
+	cp delete_folder.pl $WORKSPACE/work/$WORKING_DIR
+	cd $WORKSPACE/work/$WORKING_DIR
+	perl delete_folder.pl pom.xml
+	cp $WORKSPACE/superpom/pom.xml .
 	
-fi
+	echo "update superpom"
+	perl -pi -le 'print "<modules>" if $. == 12' pom.xml
+	perl -pi -le 'print "</modules>" if $. == 13' pom.xml
+	for s in `ls -d */ `; do
+	    repertoire=$s
+		name=`perl -e '"'$repertoire'"=~/([a-zA-Z0-9._]*)/; print $1';`
+	  	res="<module>$name</module>"
+	  	echo $repertoire $name $res
+		perl -pi -le 'print "'$res'" if $. == 13' pom.xml
+	done
+	echo "superpom patched"
+	echo "Maven deploy start"
+	mvn clean deploy -e -P public > log_maven.log
+	echo "Maven deploy done"
+	
+	sleep 5m
+	
+	tail -100 log_maven.log
+	
+	if grep "BUILD SUCCESSFUL" log_maven.log ; then
+		echo "Maven deploy success"
+	else
+	
+		if grep "ERROR" log_maven.log ; then	
+			echo "Maven deploy failed"
+			exit -2
+		else
+			echo "Maven deploy no error"
+		fi		
+	fi	
+	exit 0
 
-exit 0
+else
+   echo "Usage: delete_folder.pl must exist"
+   exit -2
+fi
